Add files via upload
This file is a version of the PowerPi app that does not contain the function to add data to the total cost column of the sqlite table. It must be run first in order to generate values for the sensor_1_rate_cost and sensor_2_rate_cost so that the total_cost function has a value to add together. Alternatively, when creating the database you can add a small value to the two relevant columns to ensure that they are not null when the PowerPi script is run.